added star shard generation
The generation is currently for every world generated loot table, i.e. all structure chest and might have to be nerved in the future. Also, it uses the [LootGenerateEvent] (https://hub.spigotmc.org/javadocs/spigot/org/bukkit/event/world/LootGenerateEvent.html) and the documentation mentions that it's "not CURRENTLY called when an entity's loot table has been generated" so the way to add the loot might also have to be changed  or modified in the future.