WrappedPlayerInteractEvent#actionPerformed
This commits adds an "actionPerformed" property to the WrappedPlayerInteractEvent that signals whether a custom action has been performed. Previously, only PlayerInteractEvent#setCancelled was used, which would not make it clear if a custom action has taken place yet.
WrappedPlayerInteractEvent is now also exposed to item behaviors and the handleInteract function will now always be called, regardless of whether a custom action has taken place or not.

Relates to #282