Add NetworkNode#isValid
The problem here was that a network node may be queued for removal during it's destruction but will continue to be part of a live network. This could theoretically lead to dupes, for example if a storage block drops its contents on destruction but the networked inventory is not cleared and still connected to an item network that ticks.