Ensure context param immutability
This commit adds a new property: ContextParamType#copy, which is a lambda to copy the param type's value. Values are copied both when they are added to the context and when they are retrieved from the context. This is intended to ensure immutability for types such as ItemStack or Location.

Specifically, this fixes an issue where TILE_ENTITY_DATA could not be read during block place because the item stack count was already decremented to 0.