Fix ProtoNetwork#isEmpty
Proto networks with only one end point can no longer be considered empty because unloaded nodes are now just removed and not replaced with ghost nodes. Even though the network only consists of one end point, it may still need to be clustered with other (local) networks, so it cannot be removed.