Fix ScaleTransform inverse display scaling not operating on effective display
The transform needs to operate on the effective display in order to also scale the default implicit display settings. Since the defaulting behavior from right to left hands is no longer encoded in model (because it is not possible to represent this correctly when parents are involved), the left hand display settings were added to the expected test results for rescaled oversized models.